fix(xxhash): uint64

+++ xxhash.go
@@ -37,10 +37,10 @@ func xxHash64(input string) uint64 {
 
 // xxHash64Large processes input data >= 32 bytes.
 func xxHash64Large(data []byte, length uint64) uint64 {
-	v1 := seed64_1
-	v2 := prime64_2
+	v1 := uint64(seed64_1)
+	v2 := uint64(prime64_2)
 	v3 := uint64(0)
-	v4 := seed64_4
+	v4 := uint64(seed64_4)
 
 	for len(data) >= 32 {
 		v1 = xxHash64Round(v1, binary.LittleEndian.Uint64(data[0:8]))
